refactor(dir): separare files and folders
