+++ data/ldifs/passport_host/gluuPassportConfiguration.json
